--- src/layouts/Footer/footer.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import styled from '@emotion/styled';
-
-const Footer: React.FCX = ({ className }) => (
-  <footer className={className}>footer</footer>
-);
-
-export const StyledFooter = styled(Footer)``;
-
-export default StyledFooter;
