+++ src/layouts/Footer/index.tsx
@@ -1,1 +1,10 @@
-export { default as Footer } from './footer';
+import React from 'react';
+import styled from '@emotion/styled';
+
+const Footer: React.FCX = ({ className }) => (
+  <footer className={className}>footer</footer>
+);
+
+export const StyledFooter = styled(Footer)``;
+
+export default StyledFooter;
