--- src/layouts/Header/header.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import styled from '@emotion/styled';
-
-const Header: React.FCX = ({ className }) => (
-  <header className={className}>header</header>
-);
-
-export const StyledHeader = styled(Header)``;
-
-export default StyledHeader;
