+++ src/layouts/Header/index.tsx
@@ -1,1 +1,10 @@
-export { default as Header } from './header';
+import React from 'react';
+import styled from '@emotion/styled';
+
+const Header: React.FCX = ({ className }) => (
+  <header className={className}>header</header>
+);
+
+export const StyledHeader = styled(Header)``;
+
+export default StyledHeader;
