+++ src/layouts/index.tsx
@@ -1,3 +1,16 @@
-export * from './layout';
-export * from './Header/header';
-export * from './Footer/footer';
+import React from 'react';
+import { ResetCSS, GlobalCSS } from '../components';
+import StyledHeader from './Header';
+import StyledFooter from './Footer';
+
+export const Layout: React.FC = ({ children }) => (
+  <>
+    <ResetCSS />
+    <GlobalCSS />
+    <StyledHeader />
+    <main>{children}</main>
+    <StyledFooter />
+  </>
+);
+
+export default Layout;
